Updated cpp keylogger, adding Python keylogger
CPP Keylogger is now working correctly.  Will need to keep messing with
it to get it to compile statically in MS so it'll be standalone, but
this shouldn't be hard.

Also found some quick code to make a Python keylogger for windows.. this
might be a more fun and faster illustration for the tutorial.